Добил одобрение товаров и пункты

+++ app/Http/Controllers/AuthController.php
@@ -67,7 +67,7 @@ class AuthController extends Controller
         if (auth('web')->attempt($data)) {
             return redirect()->route('home');
         } else {
-            return redirect()->route('sign_in')->withErrors([
+            return redirect()->route('login')->withErrors([
                 'tel' => 'Телефон или пароль не совпадают'
             ]);
         }
@@ -83,7 +83,41 @@ class AuthController extends Controller
         $user_details = UserDetails::where('user_id', '=', auth()->user()->id)->first();
         $user_details_count = UserDetails::where('user_id', '=', auth()->user()->id)->count();
         $my_friends_count = Friends::where('add_friends_id', '=', auth()->user()->id)->orWhere('app_friends_id', '=', auth()->user()->id)->count();
-        return view('profile', ['user_details' => $user_details, 'user_details_count' => $user_details_count, 'my_friends_count' => $my_friends_count]);
+        if($user_details_count != 0) {
+            if($user_details->mounth != '') {
+                $mounth_figure = $user_details->mounth;
+                if($mounth_figure == 01) {
+                    $mounth = 'января';
+                } elseif($mounth_figure == 02) {
+                    $mounth = 'февраля';
+                } elseif($mounth_figure == 03) {
+                    $mounth = 'марта';
+                } elseif($mounth_figure == 04) {
+                    $mounth = 'апреля';
+                } elseif($mounth_figure == 05) {
+                    $mounth = 'мая';
+                } elseif($mounth_figure == 06) {
+                    $mounth = 'июня';
+                } elseif($mounth_figure == 07) {
+                    $mounth = 'июля';
+                } elseif($mounth_figure == 8) {
+                    $mounth = 'августа';
+                } elseif($mounth_figure == 9) {
+                    $mounth = 'сентября';
+                } elseif($mounth_figure == 10) {
+                    $mounth = 'октября';
+                } elseif($mounth_figure == 11) {
+                    $mounth = 'ноября';
+                } elseif($mounth_figure == 12) {
+                    $mounth = 'декабря';
+                }
+                return view('profile', ['user_details' => $user_details, 'user_details_count' => $user_details_count, 'my_friends_count' => $my_friends_count, 'mounth' => $mounth]);
+            } else {
+                return view('profile', ['user_details' => $user_details, 'user_details_count' => $user_details_count, 'my_friends_count' => $my_friends_count]);
+            }
+        } else {
+            return view('profile', ['user_details' => $user_details, 'user_details_count' => $user_details_count, 'my_friends_count' => $my_friends_count]);
+        }
     }
 
     public function exit_personal() {
@@ -108,10 +142,11 @@ class AuthController extends Controller
         if($user_details == 0) {
             $user_details = new UserDetails();
             $user_details->user_id = auth()->user()->id;
-            $user_details->avatar = '';
+            $user_details->avatar = 'default.png';
             $user_details->day = $data->input('day');
             $user_details->mounth = $data->input('mounth');
             $user_details->year = $data->input('year');
+            $user_details->city = $data->input('city');
             $user_details->gender = $data->input('gender');
             $user_details->save();
         } else {
@@ -119,6 +154,7 @@ class AuthController extends Controller
             $user_details->day = $data->input('day');
             $user_details->mounth = $data->input('mounth');
             $user_details->year = $data->input('year');
+            $user_details->city = $data->input('city');
             $user_details->gender = $data->input('gender');
             $user_details->save();
         }
@@ -131,6 +167,7 @@ class AuthController extends Controller
          ]);        
 
          $user_details = UserDetails::where('user_id', '=', auth()->user()->id)->count();
+         $user_details_avatar = UserDetails::where([['user_id', '=', auth()->user()->id], ['avatar', '=', 'default.png']])->count();
 
          if($user_details == 0) {
             $user_details = new UserDetails();
@@ -144,18 +181,33 @@ class AuthController extends Controller
             $user_details->day='';
             $user_details->mounth='';
             $user_details->year='';
+            $user_details->city= '';
             $user_details->gender='';
             $user_details->save();
          } else {
-            $user_details = UserDetails::find($id);
+            $user_details = UserDetails::where('user_id', '=', $id)->first();
             $file = $data->file('avatar');
             $upload_folder = 'public/avatar/'.auth()->user()->id; //Создается автоматически
             $filename = $file->getClientOriginalName(); //Сохраняем исходное название изображения
-            Storage::delete($upload_folder.'/'.$user_details->avatar);
-            Storage::putFileAs($upload_folder, $file, $filename);
+            if($user_details_avatar == 0) {
+                Storage::delete($upload_folder.'/'.$user_details->avatar);
+            }
             $user_details->avatar=$filename;
             $user_details->save();
+            Storage::putFileAs($upload_folder, $file, $filename);
          }
          return redirect()->route('profile');
+    }
+
+    public function delete_avatar() {
+        $user_details = UserDetails::where('user_id', '=', auth()->user()->id)->first();
+
+        $upload_folder = 'public/avatar/'.auth()->user()->id;
+        Storage::delete($upload_folder.'/'.$user_details->avatar);
+
+        $user_details->avatar = 'default.png';
+        $user_details->save();
+        
+        return redirect()->route('profile');
     }
 }
